Makes changes for the latest code.
Refactor the WebHostBuilder classes to share more code.
